refactor: created `Filters` and `Notes` context
moved all filter-related code from `Personalization` to `Filters` context, and moved all notes-related code from `Personalization` to `Notes` context
