+++ lib/lanttern/personalization.ex
@@ -7,383 +7,8 @@ defmodule Lanttern.Personalization do
   alias Lanttern.Repo
   import Lanttern.RepoHelpers
 
-  alias Lanttern.Personalization.MomentNoteRelationship
-  alias Lanttern.Personalization.Note
   alias Lanttern.Personalization.ProfileSettings
-  alias Lanttern.Personalization.StrandNoteRelationship
   alias Lanttern.Personalization.ProfileView
-  alias Lanttern.Personalization.ProfileStrandFilter
-  alias Lanttern.Personalization.ProfileReportCardFilter
-  alias Lanttern.Reporting.ReportCard
-  alias Lanttern.Schools.Cycle
-  alias Lanttern.Identity.User
-  alias Lanttern.LearningContext.Moment
-  alias Lanttern.LearningContext.Strand
-
-  @doc """
-  Returns the list of notes.
-
-  ### Options:
-
-  `:preloads` – preloads associated data
-
-  ## Examples
-
-      iex> list_notes()
-      [%Note{}, ...]
-
-  """
-  def list_notes(opts \\ []) do
-    Repo.all(Note)
-    |> maybe_preload(opts)
-  end
-
-  @doc """
-  Returns the list of user notes.
-
-  ### Options (required):
-
-  `:strand_id` – list all moments notes for given strand. with preloaded moment and ordered by its position
-
-  ## Examples
-
-      iex> list_user_notes(user, opts)
-      [%Note{}, ...]
-
-  """
-  def list_user_notes(%{current_profile: profile} = _user, strand_id: strand_id) do
-    from(
-      n in Note,
-      join: mn in MomentNoteRelationship,
-      on: mn.note_id == n.id,
-      join: m in Moment,
-      on: m.id == mn.moment_id,
-      where: n.author_id == ^profile.id,
-      where: m.strand_id == ^strand_id,
-      order_by: m.position,
-      select: %{n | moment: m}
-    )
-    |> Repo.all()
-  end
-
-  @doc """
-  Returns the list of student strand notes based on their report cards.
-
-  The function lists all strands linked to the student report cards, ordering the results
-  by (report card) cycle descending and strand reports position ascending.
-
-  The list is comprised of tuples of note (or `nil`), strand, and cycle.
-
-  ### Options:
-
-  `:cycles_ids` – filter results by given cycles
-
-  ## Examples
-
-      iex> list_student_strands_notes(user, opts)
-      [{%Note{}, %Cycle{}}, ...]
-
-  """
-  @spec list_student_strands_notes(user :: User.t(), opts :: Keyword.t()) :: [
-          {Note.t() | nil, Strand.t(), Cycle.t()}
-        ]
-  def list_student_strands_notes(%{current_profile: profile} = _user, opts \\ []) do
-    notes_strands_cycles =
-      from(
-        rc in ReportCard,
-        join: c in assoc(rc, :school_cycle),
-        as: :cycles,
-        join: src in assoc(rc, :students_report_cards),
-        join: sr in assoc(rc, :strand_reports),
-        join: s in assoc(sr, :strand),
-        left_join: n in assoc(s, :notes),
-        where: src.student_id == ^profile.student_id,
-        where: is_nil(n) or n.author_id == ^profile.id,
-        order_by: [desc: c.end_at, asc: c.start_at, asc: sr.position],
-        select: {n, s, c}
-      )
-      |> apply_list_student_strands_notes_opts(opts)
-      |> Repo.all()
-
-    # preload strand subjects and years
-    updated_strands =
-      notes_strands_cycles
-      |> Enum.map(fn {_, strand, _} -> strand end)
-      |> maybe_preload(preloads: [:subjects, :years])
-
-    # put back strands with preloaded subjects and years
-    notes_strands_cycles
-    |> Enum.zip_with(
-      updated_strands,
-      fn {n, _, c}, s -> {n, s, c} end
-    )
-  end
-
-  defp apply_list_student_strands_notes_opts(queryable, []), do: queryable
-
-  defp apply_list_student_strands_notes_opts(queryable, [{:cycles_ids, cycles_ids} | opts]) do
-    from(
-      [_, cycles: c] in queryable,
-      where: c.id in ^cycles_ids
-    )
-    |> apply_list_student_strands_notes_opts(opts)
-  end
-
-  defp apply_list_student_strands_notes_opts(queryable, [_opt | opts]),
-    do: apply_list_student_strands_notes_opts(queryable, opts)
-
-  @doc """
-  Gets a single note.
-
-  Raises `Ecto.NoResultsError` if the Note does not exist.
-
-  ### Options:
-
-  `:preloads` – preloads associated data
-
-  ## Examples
-
-      iex> get_note!(123)
-      %Note{}
-
-      iex> get_note!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_note!(id, opts \\ []) do
-    Repo.get!(Note, id)
-    |> maybe_preload(opts)
-  end
-
-  @doc """
-  Gets a single user note.
-
-  Returns `nil` if the Note does not exist.
-
-  ### Options (required):
-
-  `:strand_id` – get user strand note with preloaded strand
-
-  `:moment_id` – get user moment note with preloaded moment
-
-  ## Examples
-
-      iex> get_user_note(user, opts)
-      %Note{}
-
-      iex> get_user_note(user, opts)
-      nil
-
-  """
-  def get_user_note(%{current_profile: profile} = _user, strand_id: strand_id) do
-    from(
-      n in Note,
-      join: s in assoc(n, :strand),
-      where: n.author_id == ^profile.id,
-      where: s.id == ^strand_id,
-      select: %{n | strand: s}
-    )
-    |> Repo.one()
-  end
-
-  def get_user_note(%{current_profile: profile} = _user, moment_id: moment_id) do
-    from(
-      n in Note,
-      join: m in assoc(n, :moment),
-      where: n.author_id == ^profile.id,
-      where: m.id == ^moment_id,
-      select: %{n | moment: m}
-    )
-    |> Repo.one()
-  end
-
-  @doc """
-  Gets a single student note.
-
-  Returns `nil` if the Note does not exist.
-
-  ### Options (required):
-
-  `:strand_id` – get student strand note with preloaded strand
-
-  ## Examples
-
-      iex> get_student_note(user, opts)
-      %Note{}
-
-      iex> get_student_note(user, opts)
-      nil
-
-  """
-  @spec get_student_note(student_id :: pos_integer(), Keyword.t()) :: Note.t() | nil
-  def get_student_note(student_id, strand_id: strand_id) do
-    query =
-      from(
-        n in Note,
-        # author is a profile
-        join: p in assoc(n, :author),
-        join: sn in StrandNoteRelationship,
-        on: sn.note_id == n.id,
-        join: s in Strand,
-        on: s.id == sn.strand_id,
-        where: p.student_id == ^student_id,
-        where: s.id == ^strand_id,
-        select: %{n | strand: s}
-      )
-
-    Repo.one(query)
-  end
-
-  @doc """
-  Creates a note.
-
-  ### Options:
-
-  `:preloads` – preloads associated data
-
-  ## Examples
-
-      iex> create_note(%{field: value})
-      {:ok, %Note{}}
-
-      iex> create_note(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_note(attrs \\ %{}, opts \\ []) do
-    %Note{}
-    |> Note.changeset(attrs)
-    |> Repo.insert()
-    |> maybe_preload(opts)
-  end
-
-  @doc """
-  Creates a user strand note.
-
-  ## Examples
-
-      iex> create_strand_note(user, 1, %{field: value})
-      {:ok, %Note{}}
-
-      iex> create_strand_note(user, 1, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_strand_note(%{current_profile: profile} = _user, strand_id, attrs \\ %{}) do
-    insert_query =
-      %Note{}
-      |> Note.changeset(Map.put(attrs, "author_id", profile && profile.id))
-
-    Ecto.Multi.new()
-    |> Ecto.Multi.insert(:insert_note, insert_query)
-    |> Ecto.Multi.run(
-      :link_strand,
-      fn _repo, %{insert_note: note} ->
-        %StrandNoteRelationship{}
-        |> StrandNoteRelationship.changeset(%{
-          note_id: note.id,
-          author_id: note.author_id,
-          strand_id: strand_id
-        })
-        |> Repo.insert()
-      end
-    )
-    |> Repo.transaction()
-    |> case do
-      {:error, _multi, changeset, _changes} -> {:error, changeset}
-      {:ok, %{insert_note: note}} -> {:ok, note}
-    end
-  end
-
-  @doc """
-  Creates a user moment note.
-
-  ## Examples
-
-      iex> create_moment_note(user, 1, %{field: value})
-      {:ok, %Note{}}
-
-      iex> create_moment_note(user, 1, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_moment_note(%{current_profile: profile} = _user, moment_id, attrs \\ %{}) do
-    insert_query =
-      %Note{}
-      |> Note.changeset(Map.put(attrs, "author_id", profile && profile.id))
-
-    Ecto.Multi.new()
-    |> Ecto.Multi.insert(:insert_note, insert_query)
-    |> Ecto.Multi.run(
-      :link_moment,
-      fn _repo, %{insert_note: note} ->
-        %MomentNoteRelationship{}
-        |> MomentNoteRelationship.changeset(%{
-          note_id: note.id,
-          author_id: note.author_id,
-          moment_id: moment_id
-        })
-        |> Repo.insert()
-      end
-    )
-    |> Repo.transaction()
-    |> case do
-      {:error, _multi, changeset, _changes} -> {:error, changeset}
-      {:ok, %{insert_note: note}} -> {:ok, note}
-    end
-  end
-
-  @doc """
-  Updates a note.
-
-  ### Options:
-
-  `:preloads` – preloads associated data
-
-  ## Examples
-
-      iex> update_note(note, %{field: new_value})
-      {:ok, %Note{}}
-
-      iex> update_note(note, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_note(%Note{} = note, attrs, opts \\ []) do
-    note
-    |> Note.changeset(attrs)
-    |> Repo.update()
-    |> maybe_preload(opts)
-  end
-
-  @doc """
-  Deletes a note.
-
-  ## Examples
-
-      iex> delete_note(note)
-      {:ok, %Note{}}
-
-      iex> delete_note(note)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_note(%Note{} = note) do
-    Repo.delete(note)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking note changes.
-
-  ## Examples
-
-      iex> change_note(note)
-      %Ecto.Changeset{data: %Note{}}
-
-  """
-  def change_note(%Note{} = note, attrs \\ %{}) do
-    Note.changeset(note, attrs)
-  end
 
   @doc """
   Returns the list of profile_views.
@@ -529,554 +154,4 @@ defmodule Lanttern.Personalization do
   """
   def get_profile_settings(profile_id),
     do: Repo.get_by(ProfileSettings, profile_id: profile_id)
-
-  @doc """
-  Set current profile filters.
-
-  If there's no profile setting, this function creates one.
-
-  ## Examples
-
-      iex> set_profile_current_filters(user, %{classes_ids: [1], subjects_ids: [2]})
-      {:ok, %ProfileSettings{}}
-
-      iex> set_profile_current_filters(user, %{classes_ids: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  @spec set_profile_current_filters(User.t(), attrs :: map()) ::
-          {:ok, ProfileSettings.t()} | {:error, Ecto.Changeset.t()}
-  def set_profile_current_filters(%{current_profile: %{id: profile_id}}, attrs \\ %{}),
-    do: insert_settings_or_update_filters(get_profile_settings(profile_id), profile_id, attrs)
-
-  defp insert_settings_or_update_filters(nil, profile_id, attrs) do
-    %ProfileSettings{}
-    |> ProfileSettings.changeset(%{
-      profile_id: profile_id,
-      current_filters: attrs
-    })
-    |> Repo.insert()
-  end
-
-  defp insert_settings_or_update_filters(profile_settings, _, attrs) do
-    profile_settings
-    |> ProfileSettings.changeset(%{
-      current_filters:
-        case profile_settings.current_filters do
-          nil -> attrs
-          current_filters -> Map.from_struct(current_filters) |> Map.merge(attrs)
-        end
-    })
-    |> Repo.update()
-  end
-
-  @doc """
-  Sync params with profile filters and returns the updated params.
-
-  This function updates the profile settings if there's some value in params for the given filters,
-  or sets param values based on the profile filters — effectively allowing filter persistence between
-  sessions.
-
-  Returns a tuple with `{:noop, params}` when there's no params change, or `{:updated, params}` otherwise.
-
-  ## Examples
-
-      iex> sync_params_and_profile_filters(params, user, [:classes_ids])
-      {:noop, %{"classes_ids" => ["1", "2", "3"]}}
-
-      iex> sync_params_and_profile_filters(params, user, [:classes_ids])
-      {:updated, %{"classes_ids" => ["4", "5", "6"]}}
-
-  """
-  def sync_params_and_profile_filters(
-        params,
-        %{current_profile: %{id: profile_id}} = _user,
-        filters \\ []
-      ) do
-    case get_profile_settings(profile_id) do
-      nil ->
-        # update profile current filters
-        attrs =
-          Enum.reduce(filters, %{}, fn atom_filter, attrs ->
-            str_filter = Atom.to_string(atom_filter)
-            Map.put(attrs, atom_filter, params[str_filter])
-          end)
-
-        insert_settings_or_update_filters(nil, profile_id, attrs)
-
-        # return params as is
-        {:noop, params}
-
-      profile_settings ->
-        current_filters = profile_settings.current_filters || %{}
-
-        {op, params, attrs} =
-          Enum.reduce(filters, {:noop, params, %{}}, &reduce_filter(&1, &2, current_filters))
-
-        if attrs != %{} do
-          insert_settings_or_update_filters(profile_settings, profile_id, attrs)
-        end
-
-        {op, params}
-    end
-  end
-
-  defp reduce_filter(atom_filter, {op, params, attrs}, current_filters) do
-    str_filter = Atom.to_string(atom_filter)
-
-    case {params[str_filter], Map.get(current_filters, atom_filter)} do
-      {nil, nil} ->
-        {op, params, attrs}
-
-      {nil, []} ->
-        {op, params, attrs}
-
-      {nil, filter} when is_list(filter) ->
-        params =
-          Map.put(
-            params,
-            str_filter,
-            Enum.map(filter, &"#{&1}")
-          )
-
-        {:updated, params, attrs}
-
-      {"", _filter} ->
-        attrs = Map.put(attrs, atom_filter, [])
-        {op, params, attrs}
-
-      {param, _filter} ->
-        attrs = Map.put(attrs, atom_filter, param)
-        {op, params, attrs}
-    end
-  end
-
-  @doc """
-  Returns the list of profile_strand_filters.
-
-  ## Examples
-
-      iex> list_profile_strand_filters()
-      [%ProfileStrandFilter{}, ...]
-
-  """
-  def list_profile_strand_filters() do
-    Repo.all(ProfileStrandFilter)
-  end
-
-  @doc """
-  Returns the list of current classes ids filters for the given strand and profile.
-
-  ## Examples
-
-      iex> list_profile_strand_filters_classes_ids(1, 1)
-      [1, 2, ...]
-
-  """
-  @spec list_profile_strand_filters_classes_ids(pos_integer(), pos_integer()) :: [pos_integer()]
-  def list_profile_strand_filters_classes_ids(profile_id, strand_id) do
-    from(
-      psf in ProfileStrandFilter,
-      where: psf.profile_id == ^profile_id,
-      where: psf.strand_id == ^strand_id,
-      select: psf.class_id
-    )
-    |> Repo.all()
-  end
-
-  @doc """
-  Gets a single profile_strand_filter.
-
-  Raises `Ecto.NoResultsError` if the Profile strand filter does not exist.
-
-  ## Examples
-
-      iex> get_profile_strand_filter!(123)
-      %ProfileStrandFilter{}
-
-      iex> get_profile_strand_filter!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_profile_strand_filter!(id), do: Repo.get!(ProfileStrandFilter, id)
-
-  @doc """
-  Set profile strand filters.
-
-  ## Examples
-
-      iex> set_profile_strand_filters(user, 1, %{classes_ids: [1]})
-      :ok
-
-      iex> set_profile_strand_filters(user, 1, %{classes_ids: bad_value})
-      {:error, message}
-
-  """
-  @spec set_profile_strand_filters(User.t(), pos_integer(), map()) ::
-          {:ok, any()} | {:error, any()} | Ecto.Multi.failure()
-  def set_profile_strand_filters(%{current_profile: %{id: profile_id}}, strand_id, %{
-        classes_ids: classes_ids
-      })
-      when is_list(classes_ids) do
-    # delete existing entries for given profile/strand
-    from(
-      psf in ProfileStrandFilter,
-      where: psf.profile_id == ^profile_id,
-      where: psf.strand_id == ^strand_id
-    )
-    |> Repo.delete_all()
-
-    # and insert the new values
-    base_profile_strand_filter =
-      %ProfileStrandFilter{
-        profile_id: profile_id,
-        strand_id: strand_id
-      }
-
-    Ecto.Multi.new()
-    |> multi_insert_profile_strand_filter(
-      base_profile_strand_filter,
-      classes_ids
-    )
-    |> Repo.transaction()
-  end
-
-  defp multi_insert_profile_strand_filter(multi, _base_profile_strand_filter, []), do: multi
-
-  defp multi_insert_profile_strand_filter(multi, base_profile_strand_filter, [
-         class_id | classes_ids
-       ]) do
-    %{
-      profile_id: profile_id,
-      strand_id: strand_id
-    } = base_profile_strand_filter
-
-    name = "#{profile_id}_#{strand_id}_#{class_id}"
-
-    changeset =
-      change_profile_strand_filter(
-        base_profile_strand_filter,
-        %{class_id: class_id}
-      )
-
-    multi
-    |> Ecto.Multi.insert(name, changeset)
-    |> multi_insert_profile_strand_filter(
-      base_profile_strand_filter,
-      classes_ids
-    )
-  end
-
-  @doc """
-  Creates a profile_strand_filter.
-
-  ## Examples
-
-      iex> create_profile_strand_filter(%{field: value})
-      {:ok, %ProfileStrandFilter{}}
-
-      iex> create_profile_strand_filter(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_profile_strand_filter(attrs \\ %{}) do
-    %ProfileStrandFilter{}
-    |> ProfileStrandFilter.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a profile_strand_filter.
-
-  ## Examples
-
-      iex> update_profile_strand_filter(profile_strand_filter, %{field: new_value})
-      {:ok, %ProfileStrandFilter{}}
-
-      iex> update_profile_strand_filter(profile_strand_filter, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_profile_strand_filter(%ProfileStrandFilter{} = profile_strand_filter, attrs) do
-    profile_strand_filter
-    |> ProfileStrandFilter.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a profile_strand_filter.
-
-  ## Examples
-
-      iex> delete_profile_strand_filter(profile_strand_filter)
-      {:ok, %ProfileStrandFilter{}}
-
-      iex> delete_profile_strand_filter(profile_strand_filter)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_profile_strand_filter(%ProfileStrandFilter{} = profile_strand_filter) do
-    Repo.delete(profile_strand_filter)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking profile_strand_filter changes.
-
-  ## Examples
-
-      iex> change_profile_strand_filter(profile_strand_filter)
-      %Ecto.Changeset{data: %ProfileStrandFilter{}}
-
-  """
-  def change_profile_strand_filter(%ProfileStrandFilter{} = profile_strand_filter, attrs \\ %{}) do
-    ProfileStrandFilter.changeset(profile_strand_filter, attrs)
-  end
-
-  @doc """
-  Returns the list of profile_report_card_filter.
-
-  ## Examples
-
-      iex> list_profile_report_card_filter()
-      [%ProfileReportCardFilter{}, ...]
-
-  """
-  def list_profile_report_card_filter do
-    Repo.all(ProfileReportCardFilter)
-  end
-
-  @doc """
-  Returns the list of current filters related to the the given report card and profile.
-
-  Supported filters: `classes_ids`, `linked_students_classes_ids`.
-
-  ## Examples
-
-      iex> list_profile_report_card_filters(1, 1)
-      %{classes_ids: [1, 2, ...], linked_students_classes_ids: []}
-
-  """
-  @spec list_profile_report_card_filters(
-          profile_id :: pos_integer(),
-          report_card_id :: pos_integer()
-        ) :: %{
-          classes_ids: [pos_integer()],
-          linked_students_classes_ids: [pos_integer()]
-        }
-  def list_profile_report_card_filters(profile_id, report_card_id) do
-    filters =
-      from(
-        prcf in ProfileReportCardFilter,
-        where: prcf.profile_id == ^profile_id,
-        where: prcf.report_card_id == ^report_card_id,
-        select: {prcf.class_id, prcf.linked_students_class_id}
-      )
-      |> Repo.all()
-
-    classes_ids =
-      filters
-      |> Enum.map(fn {class_id, _} -> class_id end)
-      |> Enum.filter(&Function.identity/1)
-
-    linked_students_classes_ids =
-      filters
-      |> Enum.map(fn {_, class_id} -> class_id end)
-      |> Enum.filter(&Function.identity/1)
-
-    %{
-      classes_ids: classes_ids,
-      linked_students_classes_ids: linked_students_classes_ids
-    }
-  end
-
-  @doc """
-  Gets a single profile_report_card_filter.
-
-  Raises `Ecto.NoResultsError` if the Profile report card filters does not exist.
-
-  ## Examples
-
-      iex> get_profile_report_card_filter!(123)
-      %ProfileReportCardFilter{}
-
-      iex> get_profile_report_card_filter!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_profile_report_card_filter!(id), do: Repo.get!(ProfileReportCardFilter, id)
-
-  @doc """
-  Set profile report card filters.
-
-  ## Examples
-
-      iex> set_profile_report_card_filters(user, 1, %{classes_ids: [1]})
-      :ok
-
-      iex> set_profile_report_card_filters(user, 1, %{classes_ids: bad_value})
-      {:error, message}
-
-  """
-  @spec set_profile_report_card_filters(User.t(), pos_integer(), map()) ::
-          {:ok, any()} | {:error, any()} | Ecto.Multi.failure()
-
-  def set_profile_report_card_filters(
-        %{current_profile: %{id: profile_id}},
-        report_card_id,
-        filters
-      ) do
-    # delete existing entries for given profile/report_card
-    filters
-    |> Map.keys()
-    |> Enum.each(&delete_profile_report_card_filters(&1, profile_id, report_card_id))
-
-    # and insert the new values
-    base_profile_report_card_filter =
-      %ProfileReportCardFilter{
-        profile_id: profile_id,
-        report_card_id: report_card_id
-      }
-
-    types_and_ids =
-      filters
-      |> Enum.flat_map(fn {type, ids} ->
-        Enum.map(ids, &{type, &1})
-      end)
-
-    Ecto.Multi.new()
-    |> multi_insert_profile_report_card_filter(
-      base_profile_report_card_filter,
-      types_and_ids
-    )
-    |> Repo.transaction()
-  end
-
-  defp delete_profile_report_card_filters(:classes_ids, profile_id, report_card_id) do
-    from(
-      prcf in ProfileReportCardFilter,
-      where: prcf.profile_id == ^profile_id,
-      where: prcf.report_card_id == ^report_card_id,
-      where: not is_nil(prcf.class_id)
-    )
-    |> Repo.delete_all()
-  end
-
-  defp delete_profile_report_card_filters(
-         :linked_students_classes_ids,
-         profile_id,
-         report_card_id
-       ) do
-    from(
-      prcf in ProfileReportCardFilter,
-      where: prcf.profile_id == ^profile_id,
-      where: prcf.report_card_id == ^report_card_id,
-      where: not is_nil(prcf.linked_students_class_id)
-    )
-    |> Repo.delete_all()
-  end
-
-  defp multi_insert_profile_report_card_filter(multi, _base_profile_report_card_filter, []),
-    do: multi
-
-  defp multi_insert_profile_report_card_filter(multi, base_profile_report_card_filter, [
-         {type, id} | types_and_ids
-       ]) do
-    %{
-      profile_id: profile_id,
-      report_card_id: report_card_id
-    } = base_profile_report_card_filter
-
-    # filter type to field type
-    type =
-      case type do
-        :classes_ids -> :class_id
-        :linked_students_classes_ids -> :linked_students_class_id
-      end
-
-    name = "#{profile_id}_#{report_card_id}_#{type}_#{id}"
-
-    changeset =
-      change_profile_report_card_filter(
-        base_profile_report_card_filter,
-        %{type => id}
-      )
-
-    multi
-    |> Ecto.Multi.insert(name, changeset)
-    |> multi_insert_profile_report_card_filter(
-      base_profile_report_card_filter,
-      types_and_ids
-    )
-  end
-
-  @doc """
-  Creates a profile_report_card_filter.
-
-  ## Examples
-
-      iex> create_profile_report_card_filter(%{field: value})
-      {:ok, %ProfileReportCardFilter{}}
-
-      iex> create_profile_report_card_filter(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_profile_report_card_filter(attrs \\ %{}) do
-    %ProfileReportCardFilter{}
-    |> ProfileReportCardFilter.changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a profile_report_card_filter.
-
-  ## Examples
-
-      iex> update_profile_report_card_filter(profile_report_card_filter, %{field: new_value})
-      {:ok, %ProfileReportCardFilter{}}
-
-      iex> update_profile_report_card_filter(profile_report_card_filter, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_profile_report_card_filter(
-        %ProfileReportCardFilter{} = profile_report_card_filter,
-        attrs
-      ) do
-    profile_report_card_filter
-    |> ProfileReportCardFilter.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a profile_report_card_filter.
-
-  ## Examples
-
-      iex> delete_profile_report_card_filter(profile_report_card_filter)
-      {:ok, %ProfileReportCardFilter{}}
-
-      iex> delete_profile_report_card_filter(profile_report_card_filter)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_profile_report_card_filter(%ProfileReportCardFilter{} = profile_report_card_filter) do
-    Repo.delete(profile_report_card_filter)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking profile_report_card_filter changes.
-
-  ## Examples
-
-      iex> change_profile_report_card_filter(profile_report_card_filter)
-      %Ecto.Changeset{data: %ProfileReportCardFilter{}}
-
-  """
-  def change_profile_report_card_filter(
-        %ProfileReportCardFilter{} = profile_report_card_filter,
-        attrs \\ %{}
-      ) do
-    ProfileReportCardFilter.changeset(profile_report_card_filter, attrs)
-  end
 end
