+++ lib/lanttern_web/live/pages/curriculum/component/id/curriculum_component_live.html.heex
@@ -154,7 +154,7 @@
   <.markdown text={@curriculum_component.description} />
 </section>
 <.live_component
-  module={LantternWeb.Personalization.FiltersOverlayComponent}
+  module={LantternWeb.Filters.FiltersOverlayComponent}
   id="subjects-filter-modal"
   current_user={@current_user}
   title={gettext("Filter curriculum items by subject")}
@@ -162,7 +162,7 @@
   navigate={~p"/curriculum/component/#{@curriculum_component}"}
 />
 <.live_component
-  module={LantternWeb.Personalization.FiltersOverlayComponent}
+  module={LantternWeb.Filters.FiltersOverlayComponent}
   id="years-filter-modal"
   current_user={@current_user}
   title={gettext("Filter curriculum items by year")}
