+++ lib/lanttern_web/live/pages/strands/strands_live.html.heex
@@ -96,7 +96,7 @@
   </:actions>
 </.slide_over>
 <.live_component
-  module={LantternWeb.Personalization.FiltersOverlayComponent}
+  module={LantternWeb.Filters.FiltersOverlayComponent}
   id="strands-year-filters"
   current_user={@current_user}
   title={gettext("Strands year filters")}
@@ -104,7 +104,7 @@
   navigate={~p"/strands"}
 />
 <.live_component
-  module={LantternWeb.Personalization.FiltersOverlayComponent}
+  module={LantternWeb.Filters.FiltersOverlayComponent}
   id="strands-subject-filters"
   current_user={@current_user}
   title={gettext("Strands subject filters")}
