+++ lib/lanttern_web/live/shared/filters/inline_filters_component.ex
@@ -1,7 +1,7 @@
-defmodule LantternWeb.Personalization.InlineFiltersComponent do
+defmodule LantternWeb.Filters.InlineFiltersComponent do
   @moduledoc """
   Renders an inline filter component based on the list of items and selected ids
-  from `LantternWeb.PersonalizationHelpers.assign_user_filters/4`.
+  from `LantternWeb.FiltersHelpers.assign_user_filters/4`.
 
   This component receives the inital state from the parent view/component,
   but handles its own internal state (selected items).
